fix: make the remote client stack work without fastapi installed
fastapi only ships with the os extra, so a bare pip install agno must be
able to instantiate RemoteAgent, RemoteTeam, and RemoteWorkflow. Each of
them builds an AgentOSClient in __init__, and importing agno.client.os
pulled fastapi in twice over: its own module-level UploadFile import, and
the response schema modules under agno.os, whose package inits eagerly
import the FastAPI app and routers.

- agno/client/os.py: import UploadFile under TYPE_CHECKING with a runtime
  Any binding so get_type_hints() and Function.from_callable() keep
  resolving the upload annotations, and route the multipart branch on
  agno.media.File instead of UploadFile (isinstance against Any raises).
- agno/remote/base.py: add the same runtime binding for its TYPE_CHECKING
  UploadFile import, which get_type_hints() could not resolve even with
  fastapi installed.
- agno/os/__init__.py and the ten router package inits: resolve the app
  and router factories lazily via module __getattr__ so importing the
  pure-pydantic schema modules no longer executes fastapi imports.
- agno/os/schema_utils.py: new fastapi-free home for the eight dict-to-
  schema helpers the schema modules need at client parse time
  (to_utc_datetime, get_run_input, get_session_name, extract_input_media,
  stringify_input_content, format_duration_ms, format_tools,
  format_team_tools), moved verbatim from agno/os/utils.py, which
  re-exports them for existing callers.

Regression tests run the no-fastapi cases in a subprocess with fastapi
masked, assert import agno.client stays fastapi-free, and pin the
multipart payload for both agno File and fastapi UploadFile. The
generated tool schema for upload_knowledge_content is byte-identical to
the previous behavior with fastapi installed.
