+++ libs/agno/agno/os/routers/__init__.py
@@ -1,3 +1,17 @@
-from agno.os.routers.session.session import get_session_router
+from typing import TYPE_CHECKING, Any
+
+if TYPE_CHECKING:
+    from agno.os.routers.session.session import get_session_router
 
 __all__ = ["get_session_router"]
+
+
+def __getattr__(name: str) -> Any:
+    # Lazy so the pure-pydantic schema modules in this package stay importable
+    # without the `os` extra (fastapi); the router is only pulled in when used.
+    if name == "get_session_router":
+        from agno.os.routers.session.session import get_session_router
+
+        globals()[name] = get_session_router
+        return get_session_router
+    raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
